Optimizations

+++ src/renderer/RenderTarget.h
@@ -21,7 +21,7 @@ public:
 
     glm::u32vec2& GetSize() { return m_Size; }
     void SetSize(const glm::u32vec2 &size);
-    void SetSize(const unsigned int &width, const unsigned int &height);
+    void SetSize(const size_t &width, const size_t &height);
 
     void BindTarget(const bool &clearColor, const bool &clearDepth);
     void UnbindTarget();
